+++ src/Middleware/RequestThrottling/src/RequestThrottlingOptions.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.Aspnetcore.RequestThrottling;
+using Microsoft.AspNetCore.RequestThrottling;
 
 namespace Microsoft.AspNetCore.RequestThrottling
 {
